moving Connect screen to it's own component
